[PRDispatcher] WIP python/workers: make diff_tools_shared module for diffbrowsers and diffenator.
